+++ src/core/config/usage.h
@@ -136,6 +136,7 @@ static inline const std::string &usage()
     u += "      --health-print-time=N     print health report every N seconds\n";
 #   endif
     u += "      --no-color                disable colored output\n";
+    u += "      --verbose                 verbose output\n";
 
     u += "\nMisc:\n";