+++ src/crypto/common/Algorithm.cpp
@@ -112,6 +112,10 @@ static AlgoName const algorithm_names[] = {
     { "RandomXL",                  nullptr,            Algorithm::RX_LOKI         },
     { "randomx/arq",               "rx/arq",           Algorithm::RX_ARQ          },
     { "RandomARQ",                 nullptr,            Algorithm::RX_ARQ          },
+    { "randomx/sfx",               "rx/sfx",           Algorithm::RX_SFX          },
+    { "RandomSFX",                 nullptr,            Algorithm::RX_SFX          },
+    { "randomx/v",                 "rx/v",             Algorithm::RX_V            },
+    { "RandomV",                   nullptr,            Algorithm::RX_V            },
 #   endif
 #   ifdef XMRIG_ALGO_ARGON2
     { "argon2/chukwa",             nullptr,            Algorithm::AR2_CHUKWA      },
@@ -138,6 +142,8 @@ size_t xmrig::Algorithm::l2() const
     switch (m_id) {
     case RX_0:
     case RX_LOKI:
+    case RX_SFX:
+    case RX_V:
         return 0x40000;
 
     case RX_WOW:
@@ -173,6 +179,8 @@ size_t xmrig::Algorithm::l3() const
         switch (m_id) {
         case RX_0:
         case RX_LOKI:
+        case RX_SFX:
+        case RX_V:
             return oneMiB * 2;
 
         case RX_WOW:
@@ -272,6 +280,8 @@ xmrig::Algorithm::Family xmrig::Algorithm::family(Id id)
     case RX_WOW:
     case RX_LOKI:
     case RX_ARQ:
+    case RX_SFX:
+    case RX_V:
         return RANDOM_X;
 #   endif