+++ src/crypto/rx/RxAlgo.cpp
@@ -49,6 +49,12 @@ const RandomX_ConfigurationBase *xmrig::RxAlgo::base(Algorithm::Id algorithm)
     case Algorithm::RX_ARQ:
         return &RandomX_ArqmaConfig;
 
+    case Algorithm::RX_SFX:
+        return &RandomX_SafexConfig;
+
+    case Algorithm::RX_V:
+        return &RandomX_VConfig;
+
     default:
         break;
     }